Rename SendAnalytics file
