+++ packages/design-system/src/components/HelpDrawer/HelpDrawer.tsx
@@ -1,4 +1,4 @@
-import { EVENT_CATEGORY, MAX_LENGTH, sendLinkEvent } from '../analytics/SendAnalytics';
+import { EVENT_CATEGORY, MAX_LENGTH, sendLinkEvent } from '../analytics';
 import React from 'react';
 import Drawer, { DrawerProps } from '../Drawer/Drawer';
 import { helpDrawerSendsAnalytics } from '../flags';
