+++ packages/design-system/src/components/analytics/index.ts
@@ -1,2 +1,1 @@
-export { default as sendAnalytics, sendLinkEvent } from './SendAnalytics';
-export type { UtagContainer, UtagObject } from './SendAnalytics';
+export * from './events';
